requirements.txt
diffusers==0.12.1
transformers
salesforce-lavis
openai-clip
accelerate
Pillow==9.1.0
einops